further abstractions
